pyproject.toml
[project]
name = "icy-disk"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "easyocr>=1.7.2",
    "fastapi>=0.135.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
